scyllaclient: column_family_total_disk_space_used returns int64
Fixes #1357

+++ scyllaclient/client_scylla.go
@@ -480,5 +480,5 @@ func (c *Client) TableDiskSize(ctx context.Context, host, keyspace, table string
 	if err != nil {
 		return 0, err
 	}
-	return int64(resp.Payload), nil
+	return resp.Payload, nil
 }
